backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
PyJWT[crypto]==2.9.0
httpx==0.27.0
pydantic==2.8.2
SQLAlchemy==2.0.32
pg8000==1.31.2
python-dotenv==1.0.1
